Change tests to work with different source.ids to avoid sync conflicts
