+++ code/test/Core.Test/Properties/AssemblyInfo.cs
@@ -42,4 +42,4 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyVersion("1.0.0.0")]
 [assembly: AssemblyFileVersion("0.0.0.0")]
 
-[assembly: Xunit.CollectionBehavior(DisableTestParallelization = true)]
+//[assembly: Xunit.CollectionBehavior(DisableTestParallelization = true)]
