+++ code/test/Templates.Test/BaseGenAndBuildTests.cs
@@ -22,7 +22,7 @@ namespace Microsoft.Templates.Test
 {
     public class BaseGenAndBuildTests : BaseTestContextProvider
     {
-        protected GenerationFixture _fixture;
+        protected BaseGenAndBuildFixture _fixture;
 
         private async Task SetUpFixtureForTestingAsync(string language)
         {
@@ -33,17 +33,18 @@ namespace Microsoft.Templates.Test
         {
             await SetUpFixtureForTestingAsync(language);
 
-            var targetProjectTemplate = GenerationFixture.Templates.FirstOrDefault(projectTemplateSelector);
+            var targetProjectTemplate = _fixture.Templates().FirstOrDefault(projectTemplateSelector);
 
             ProjectName = projectName;
+
             ProjectPath = Path.Combine(_fixture.TestProjectsPath, projectName, projectName);
             OutputPath = ProjectPath;
 
-            var userSelection = await GenerationFixture.SetupProjectAsync(projectType, framework, language, getName);
+            var userSelection = _fixture.SetupProject(projectType, framework, language, getName);
 
             if (getName != null)
             {
-                GenerationFixture.AddItems(userSelection, GenerationFixture.GetTemplates(framework), getName);
+                _fixture.AddItems(userSelection, _fixture.GetTemplates(framework), getName);
             }
 
             await NewProjectGenController.Instance.UnsafeGenerateProjectAsync(userSelection);
@@ -66,132 +67,132 @@ namespace Microsoft.Templates.Test
         protected void AssertBuildProjectAsync(string projectPath, string projectName)
         {
             // Build solution
-            var result = GenerationFixture.BuildSolution(projectName, projectPath);
+            var result = _fixture.BuildSolution(projectName, projectPath);
 
             // Assert
-            Assert.True(result.exitCode.Equals(0), $"Solution {projectName} was not built successfully. {Environment.NewLine}Errors found: {GenerationFixture.GetErrorLines(result.outputFile)}.{Environment.NewLine}Please see {Path.GetFullPath(result.outputFile)} for more details.");
+            Assert.True(result.exitCode.Equals(0), $"Solution {projectName} was not built successfully. {Environment.NewLine}Errors found: {_fixture.GetErrorLines(result.outputFile)}.{Environment.NewLine}Please see {Path.GetFullPath(result.outputFile)} for more details.");
 
             // Clean
             Fs.SafeDeleteDirectory(projectPath);
         }
 
-        protected async Task<string> AssertGenerateRightClickAsync(string projectName, string projectType, string framework, string language, bool emptyProject, bool cleanGeneration = true)
-        {
-            await SetUpFixtureForTestingAsync(language);
+        //protected async Task<string> AssertGenerateRightClickAsync(string projectName, string projectType, string framework, string language, bool emptyProject, bool cleanGeneration = true)
+        //{
+        //    await SetUpFixtureForTestingAsync(language);
 
-            ProjectName = projectName;
-            ProjectPath = Path.Combine(_fixture.TestNewItemPath, projectName, projectName);
-            OutputPath = ProjectPath;
+        //    ProjectName = projectName;
+        //    ProjectPath = Path.Combine(_fixture.TestNewItemPath, projectName, projectName);
+        //    OutputPath = ProjectPath;
 
-            var userSelection = await GenerationFixture.SetupProjectAsync(projectType, framework, language);
+        //    var userSelection = await _fixture.SetupProjectAsync(projectType, framework, language);
 
-            if (!emptyProject)
-            {
-                GenerationFixture.AddItems(userSelection, GenerationFixture.GetTemplates(framework), GenerationFixture.GetDefaultName);
-            }
+        //    if (!emptyProject)
+        //    {
+        //        _fixture.AddItems(userSelection, _fixture.GetTemplates(framework), GenerationFixture.GetDefaultName);
+        //    }
 
-            await NewProjectGenController.Instance.UnsafeGenerateProjectAsync(userSelection);
+        //    await NewProjectGenController.Instance.UnsafeGenerateProjectAsync(userSelection);
 
-            var project = Path.Combine(_fixture.TestNewItemPath, projectName);
+        //    var project = Path.Combine(_fixture.TestNewItemPath, projectName);
 
-            // Assert on project
-            Assert.True(Directory.Exists(project));
+        //    // Assert on project
+        //    Assert.True(Directory.Exists(project));
 
-            int emptyProjecFileCount = Directory.GetFiles(project, "*.*", SearchOption.AllDirectories).Count();
-            Assert.True(emptyProjecFileCount > 2);
+        //    int emptyProjecFileCount = Directory.GetFiles(project, "*.*", SearchOption.AllDirectories).Count();
+        //    Assert.True(emptyProjecFileCount > 2);
 
-            // Add new items
-            var rightClickTemplates = GenerationFixture.Templates.Where(
-                                            t => (t.GetTemplateType() == TemplateType.Feature || t.GetTemplateType() == TemplateType.Page)
-                                              && t.GetFrameworkList().Contains(framework)
-                                              && !t.GetIsHidden()
-                                              && t.GetRightClickEnabled());
+        //    // Add new items
+        //    var rightClickTemplates = _fixture.Templates().Where(
+        //                                    t => (t.GetTemplateType() == TemplateType.Feature || t.GetTemplateType() == TemplateType.Page)
+        //                                      && t.GetFrameworkList().Contains(framework)
+        //                                      && !t.GetIsHidden()
+        //                                      && t.GetRightClickEnabled());
 
-            foreach (var item in rightClickTemplates)
-            {
-                OutputPath = GenContext.GetTempGenerationPath(projectName);
+        //    foreach (var item in rightClickTemplates)
+        //    {
+        //        OutputPath = GenContext.GetTempGenerationPath(projectName);
 
-                var newUserSelection = new UserSelection
-                {
-                    ProjectType = projectType,
-                    Framework = framework,
-                    HomeName = "",
-                    Language = language,
-                    ItemGenerationType = ItemGenerationType.GenerateAndMerge
-                };
+        //        var newUserSelection = new UserSelection
+        //        {
+        //            ProjectType = projectType,
+        //            Framework = framework,
+        //            HomeName = "",
+        //            Language = language,
+        //            ItemGenerationType = ItemGenerationType.GenerateAndMerge
+        //        };
 
-                GenerationFixture.AddItem(newUserSelection, item, GenerationFixture.GetDefaultName);
+        //        _fixture.AddItem(newUserSelection, item, GenerationFixture.GetDefaultName);
 
-                await NewItemGenController.Instance.UnsafeGenerateNewItemAsync(item.GetTemplateType(), newUserSelection);
+        //        await NewItemGenController.Instance.UnsafeGenerateNewItemAsync(item.GetTemplateType(), newUserSelection);
 
-                NewItemGenController.Instance.UnsafeFinishGeneration(newUserSelection);
-            }
+        //        NewItemGenController.Instance.UnsafeFinishGeneration(newUserSelection);
+        //    }
 
-            var finalProjectPath = Path.Combine(_fixture.TestNewItemPath, projectName);
-            int finalProjectFileCount = Directory.GetFiles(finalProjectPath, "*.*", SearchOption.AllDirectories).Count();
+        //    var finalProjectPath = Path.Combine(_fixture.TestNewItemPath, projectName);
+        //    int finalProjectFileCount = Directory.GetFiles(finalProjectPath, "*.*", SearchOption.AllDirectories).Count();
 
-            if (emptyProject)
-            {
-                Assert.True(finalProjectFileCount > emptyProjecFileCount);
-            }
-            else
-            {
-                Assert.True(finalProjectFileCount == emptyProjecFileCount);
-            }
-            // Clean
-            if (cleanGeneration)
-            {
-                Fs.SafeDeleteDirectory(finalProjectPath);
-            }
+        //    if (emptyProject)
+        //    {
+        //        Assert.True(finalProjectFileCount > emptyProjecFileCount);
+        //    }
+        //    else
+        //    {
+        //        Assert.True(finalProjectFileCount == emptyProjecFileCount);
+        //    }
+        //    // Clean
+        //    if (cleanGeneration)
+        //    {
+        //        Fs.SafeDeleteDirectory(finalProjectPath);
+        //    }
 
-            return finalProjectPath;
-        }
-        protected async Task<(string ProjectPath, string ProjecName)> AssertGenerationOneByOneAsync(string itemName, string projectType, string framework, string itemId, string language, bool cleanGeneration = true)
-        {
-            await SetUpFixtureForTestingAsync(language);
+        //    return finalProjectPath;
+        //}
+        //protected async Task<(string ProjectPath, string ProjecName)> AssertGenerationOneByOneAsync(string itemName, string projectType, string framework, string itemId, string language, bool cleanGeneration = true)
+        //{
+        //    await SetUpFixtureForTestingAsync(language);
 
-            var projectTemplate = GenerationFixture.Templates.FirstOrDefault(t => t.GetTemplateType() == TemplateType.Project && t.GetProjectTypeList().Contains(projectType) && t.GetFrameworkList().Contains(framework));
-            var itemTemplate = GenerationFixture.Templates.FirstOrDefault(t => t.Identity == itemId);
-            var finalName = itemTemplate.GetDefaultName();
-            var validators = new List<Validator>
-            {
-                new ReservedNamesValidator(),
-            };
-            if (itemTemplate.GetItemNameEditable())
-            {
-                validators.Add(new DefaultNamesValidator());
-            }
+        //    var projectTemplate = _fixture.Templates().FirstOrDefault(t => t.GetTemplateType() == TemplateType.Project && t.GetProjectTypeList().Contains(projectType) && t.GetFrameworkList().Contains(framework));
+        //    var itemTemplate = _fixture.Templates().FirstOrDefault(t => t.Identity == itemId);
+        //    var finalName = itemTemplate.GetDefaultName();
+        //    var validators = new List<Validator>
+        //    {
+        //        new ReservedNamesValidator(),
+        //    };
+        //    if (itemTemplate.GetItemNameEditable())
+        //    {
+        //        validators.Add(new DefaultNamesValidator());
+        //    }
 
-            finalName = Naming.Infer(finalName, validators);
+        //    finalName = Naming.Infer(finalName, validators);
 
-            var projectName = $"{projectType}{framework}{finalName}";
+        //    var projectName = $"{projectType}{framework}{finalName}";
 
-            ProjectName = projectName;
-            ProjectPath = Path.Combine(_fixture.TestProjectsPath, projectName, projectName);
-            OutputPath = ProjectPath;
+        //    ProjectName = projectName;
+        //    ProjectPath = Path.Combine(_fixture.TestProjectsPath, projectName, projectName);
+        //    OutputPath = ProjectPath;
 
-            var userSelection = await GenerationFixture.SetupProjectAsync(projectType, framework, language);
+        //    var userSelection = await _fixture.SetupProjectAsync(projectType, framework, language);
 
-            GenerationFixture.AddItem(userSelection, itemTemplate, GenerationFixture.GetDefaultName);
+        //    _fixture.AddItem(userSelection, itemTemplate, GenerationFixture.GetDefaultName);
 
-            await NewProjectGenController.Instance.UnsafeGenerateProjectAsync(userSelection);
+        //    await NewProjectGenController.Instance.UnsafeGenerateProjectAsync(userSelection);
 
-            var resultPath = Path.Combine(_fixture.TestProjectsPath, projectName);
+        //    var resultPath = Path.Combine(_fixture.TestProjectsPath, projectName);
 
-            // Assert
-            Assert.True(Directory.Exists(resultPath));
-            Assert.True(Directory.GetFiles(resultPath, "*.*", SearchOption.AllDirectories).Count() > 2);
+        //    // Assert
+        //    Assert.True(Directory.Exists(resultPath));
+        //    Assert.True(Directory.GetFiles(resultPath, "*.*", SearchOption.AllDirectories).Count() > 2);
 
-            // Clean
-            if (cleanGeneration)
-            {
-                Fs.SafeDeleteDirectory(resultPath);
-            }
+        //    // Clean
+        //    if (cleanGeneration)
+        //    {
+        //        Fs.SafeDeleteDirectory(resultPath);
+        //    }
 
-            return (resultPath, projectName);
-        }
+        //    return (resultPath, projectName);
+        //}
 
-        public static IEnumerable<object[]> GetProjectTemplatesAsync()
+        public static IEnumerable<object[]> GetProjectTemplatesForGenerationAsync()
         {
             JoinableTaskContext context = new JoinableTaskContext();
             JoinableTaskCollection tasks = context.CreateCollection();
@@ -201,13 +202,23 @@ namespace Microsoft.Templates.Test
             return result;
         }
 
-        public static IEnumerable<object[]> GetPageAndFeatureTemplatesAsync(string framework)
+        public static IEnumerable<object[]> GetProjectTemplatesForBuildAsync()
         {
             JoinableTaskContext context = new JoinableTaskContext();
             JoinableTaskCollection tasks = context.CreateCollection();
             context.CreateFactory(tasks);
-            var result = context.Factory.Run(() => GenerationFixture.GetPageAndFeatureTemplatesAsync(framework));
+            var result = context.Factory.Run(() => BuildFixture.GetProjectTemplatesAsync());
+
             return result;
         }
+
+        //public static IEnumerable<object[]> GetPageAndFeatureTemplatesAsync(string framework)
+        //{
+        //    JoinableTaskContext context = new JoinableTaskContext();
+        //    JoinableTaskCollection tasks = context.CreateCollection();
+        //    context.CreateFactory(tasks);
+        //    var result = context.Factory.Run(() => GenerationFixture.GetPageAndFeatureTemplatesAsync(framework));
+        //    return result;
+        //}
     }
 }
