+++ code/test/Templates.Test/Templates.Test.csproj
@@ -120,18 +120,22 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="BuildProjectsTests.cs" />
-    <Compile Include="GenProjectsTests.cs" />
+    <Compile Include="BuildTests\BuildProjectsTests.cs" />
+    <Compile Include="BuildTests\BuildFixture.cs" />
+    <Compile Include="GenTests\GenTestsTemplateSource.cs" />
+    <Compile Include="BuildTests\BuildTestsTemplateSource.cs" />
+    <Compile Include="GenTests\GenProjectsTests.cs" />
     <Compile Include="FileEncodingTests.cs" />
     <Compile Include="BaseTestContextProvider.cs" />
     <Compile Include="CodeStyleEnforcementTests.cs" />
     <Compile Include="GlobalSuppressions.cs" />
     <Compile Include="BaseGenAndBuildTests.cs" />
+    <Compile Include="BaseGenAndBuildFixture.cs" />
     <Compile Include="StyleCopCollection.cs" />
     <Compile Include="StyleCopGenerationTestsFixture.cs" />
     <Compile Include="StyleCopProjectGenerationTests.cs" />
     <Compile Include="GenerationCollection.cs" />
-    <Compile Include="GenerationFixture.cs" />
+    <Compile Include="GenTests\GenerationFixture.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="StyleCopPlusLocalTemplatesSource.cs" />
     <AdditionalFiles Include="..\..\StyleCop.json">
